set spawnpoint manually when player sleeps
When the player sleeps, the spawnpoint is reset. This is, however,
dependent on a bed being present at the location of the spawnpoint.

This means that if the bed breaks after the player sleeps, any respawn
would go back to the default: the death chamber. This means that
sleeping in bed, then breaking a bed, then travelling to The End, would
result in the player returning to the overworld in the death chamber,
rather than anywhere useful.

To avoid this, reset the spawnpoint (manually) whenever the player
sleeps.

This solution isn't perfect. The player can still set their spawnpoint
by right-clicking a bed *without* sleeping in it, which circumvents this
attempted fix. Hopefully, it will reduce the problem.