dragon fight for everyone
when the player dies after defeating the dragon, they should still get
a new dragon fight when they make it back to the end.

This is complicated by the fact that there is no fully automatic way to
trigger the respawn sequence. However, we *can* do most of the work,
and then have the player do the last step.