bugfix: only remove "init" tags from non-players
when a new player joined during initialization of another player,
the "initialization-related" tags of both players would be cleared.
This wasn't much of an issue, as it would simply delay initialization
of the original player for another cycle.

but it was always intended to only remove the initialization-related
tags from the players who were about to be initialized.