refactor initialisation to use global scoreboards
to prevent edge-cases involving the marker not being detected, we move
the game "state" determination into a global variable.

As part of this change, the "restart required" state is now skipped if
we detect that the server version is 1.14+, as this version updates
compass direction and chunkloading without a restart or player.