rewrite dragon respawn for portal/dragon detection
there were various issues surrounding portal and dragon detection,
meaning the detection of the dragon (and the respawning of the dragon)
were both broken.

We fix this by first fixing the portal detection, looking at multiple
y-levels for the portal. Then, we use detection of the portal's state
as a fallback for detection of the dragon itself.

While we use detection of the dragon itself as a "fast" indicator of the
dragon's state, we don't decide that the dragon definitely *doesn't*
exist until we have also detected an open portal.

This new portal detection assumes that the bedrock above the portal is
not broken, which is perhaps not a good assumption.