upgrade path
Create an upgrade path for previous versions of foreverworld.
Backups should still be created, of course, but in theory this should
allow the new version of foreverworld to run without accidentally
resetting the player.

The "version" is a simple increment, and not tied to the release number
in any way, other than being intended to be added enough to ensure
compatibility between major releases.