rewrite player spawnpoint detection
as of 1.15, the player no longer needs to sleep to set their spawnpoint.
Our previous spawnpoint detection code used "sleep" as an indicator that
the spawnpoint needed to be re-checked, so this code was no-longer
valid.

Instead, add a more-direct check for "forced spawning", and immediately
re-set the spawnpoint any time we find it is not true. This simplifies
the spawnpoint code quite a bit, and so would have probably been a good
change even prior to 1.15.

This leaves us with the same issue we previously had - a lack of bed
detection / lack of fallback for when the resulting spawnpoint is not
"safe", but potentially begins a step in the right direction for
correcting this. That is: the code we use here is closely related to the
code we had intended to use to allow for "bed breaking" to return the
player to their original-per-life spawnpoint. The full code for that
system, if it ever works fully, will be added in a separate commit.