do not force spawnpoints
previously, we worked-around the issue of "broken beds returning the
player to the death chamber" by forcing all spawnpoints. ie, when the
player set their spawnpoint, we would use the /spawnpoint command to
force their spawnpoint to be exactly there, regardless of whether the
bed was broken.

This was less-similar to "regular" hardcore than we wanted. Ideally, if
the bed was broken, we would want to return the player to their
"first, for this life" starting point.

This commit attempts to correct this. Instead of setting the spawnpoint,
we detect whenever a "living" player is in the death chamber. When this
occurs, we assume that the player has reached the death chamber by
leaving The End with a broken bed, and attempt to teleport them to their
original coordinates (ie, those reached via /spreadplayers when they
last died)

This should not impact "new" players, as this should only detect players
who have already been through their initialization.