fix dragon warning timings
the message "the dragon is coming" lasted much longer than it should have,
meaning that the player was left with the message on their screen even
after the dragon fight had theoretically already begun.

reduce the "stay" length of the message, so it goes away almost
instantly when the dragon is seen.

we don't remove the message "immediately", because having the message
last for only one or two ticks can cause it to flicker. We could
forcibly clear the message when we detect that the dragon has been seen,
but we don't want to interfere with potential unrelated messages.