legacy spawnpoint handling
when we added the new spawnpoint handling, we neglected to handle
existing worlds, which did not previously have the initial spawnpoints
stored. As a work-around, we set the initial spawnpoint to the value of
the player's "current" spawnpoint, which would either be their initial
spawnpoint, or the bed they last slept in. This is not a perfect
solution, but is really the best we can do for now.