force spreadplayers into a circle, not a square
The "spreadplayers" function spreads players in a square, not a circle,
while Strongholds generate in a circular pattern. Due to this, the
"corners" of the square can sometimes be exceedingly far away from any
stronghold (up to 12916 blocks away from the maximum Stronghold spawning
radius). To avoid this, we check this distance from 0 0 after spreading,
and re-spread when we are outside of the desired radius.

Due to this tweak, we can be more-specific about our spawning radius:
instead of 24000 (which was a compromise to help "cut the corners" of
the circle we actually wanted), we instead spawn in a radius of 26240,
ie: 24192 (the maximum distance from 0 0 that a stronghold can generate,
plus the width of a fully-zoomed-out map)