make lodestone compasses point north
this is why we didn't have a 1.16 release - now that there's an "item
modify" command, we can easily make lodestone compasses point north,
too.

There are a couple of caveats here:

 - it needs to run for every inventory slot
 - it only modifies compasses within inventory slots
 - it needs to delay for 1 tick in order to be detected by the client

combined, this means the rule can be bypassed by, for example, a player
with a full inventory clicking on a lodestone which is directly above a
hopper minecart - this would pull the lodestone into the hopper's
inventory, and that inventory is not checked.

This bug is currently seen as acceptably convoluted soas to not be an
issue, though it may need to be revisited in future as one could
potentially use this to get a lodestone compass into a llama's chest,
creating a mobile inventory which contains a lodestone compass.