move the chamber to the new lowest height
the death chamber is meant to extend to the bottom of the world. The
bottom of the world has moved.

the death chamber is meant to extend to the top of the world. the top of
the world has also moved.

resize the death chamber, and move the positions in which we check for
the player within it.